+++ CustomCommands/Events/DebugTests.cs
@@ -3,6 +3,7 @@ using InventorySystem.Items.Firearms.Attachments;
 using MapGeneration;
 using Mirror;
 using PlayerRoles.PlayableScps.Scp079;
+using PlayerStatsSystem;
 using PluginAPI.Core;
 using PluginAPI.Core.Attributes;
 using PluginAPI.Enums;
@@ -16,24 +17,24 @@ using UnityEngine;
 
 namespace CustomCommands.Events
 {
-	public class DebugTests
-	{
-		[PluginEvent(ServerEventType.PlayerDamagedShootingTarget)]
-		public void TargetDamagedEvent(PlayerDamagedShootingTargetEvent args)
-		{
-			if (args.Player.CurrentItem is Firearm firearm)
-			{
-				Log.Info($"{firearm.GetCurrentAttachmentsCode()}");
-			}
-		}
+    public class DebugTests
+    {
+        [PluginEvent(ServerEventType.PlayerDamagedShootingTarget)]
+        public void TargetDamagedEvent(PlayerDamagedShootingTargetEvent args)
+        {
+            if (args.Player.CurrentItem is Firearm firearm)
+            {
+                Log.Info($"{firearm.GetCurrentAttachmentsCode()}");
+            }
+        }
 
-		//[PluginEvent(ServerEventType.MapGenerated)]
-		public void MapGenEvent(MapGeneratedEvent args)
-		{
-			foreach(var a in NetworkClient.prefabs)
-			{
-				Log.Info($"{a.Value.name}");
-			}
-		}
-	}
+        //[PluginEvent(ServerEventType.MapGenerated)]
+        public void MapGenEvent(MapGeneratedEvent args)
+        {
+            foreach (var a in NetworkClient.prefabs)
+            {
+                Log.Info($"{a.Value.name}");
+            }
+        }
+    }
 }
