+++ CustomCommands/Events/SCP3114Overhaul.cs
@@ -6,11 +6,6 @@ using PlayerRoles;
 using PluginAPI.Core.Attributes;
 using PluginAPI.Enums;
 using PluginAPI.Events;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using UnityEngine;
 using PluginAPI.Core;
 using MEC;
