+++ CustomCommands/Plugin.cs
@@ -63,10 +63,10 @@ namespace CustomCommands
 			EventManager.RegisterEvents<LateJoin>(this);
 			EventManager.RegisterEvents<NameFix>(this);
 			EventManager.RegisterEvents<SCPDamageAnnouncement>(this);		
-			EventManager.RegisterEvents<SCPSwap>(this);
+			EventManager.RegisterEvents<Features.SCPSwap>(this);
 			EventManager.RegisterEvents<SurfaceLightingFix>(this);
 			EventManager.RegisterEvents<TutorialFixes>(this);
-			EventManager.RegisterEvents<Voting>(this);
+			EventManager.RegisterEvents<Features.Voting>(this);
 			//EventManager.RegisterEvents<SCP3114Overhaul>(this);
 
 			//EventManager.RegisterEvents<SCP008>(this);
@@ -75,16 +75,6 @@ namespace CustomCommands
 			//InventoryExtensions.OnItemRemoved += InventoryExtensions_OnItemRemoved;
 
 			//RagdollManager.OnRagdollSpawned += RagdollManager_OnRagdollSpawned;
-
-			CharacterClassManager.OnRoundStarted += CharacterClassManager_OnRoundStarted;
-		}
-
-		private void CharacterClassManager_OnRoundStarted()
-		{
-			Timing.CallDelayed(0.5f, () =>
-			{
-				GameObject.FindObjectOfType<SkyboxHubert>().NetworkHubert = true;
-			});
 		}
 
 		private void InventoryExtensions_OnItemRemoved(ReferenceHub refHub, InventorySystem.Items.ItemBase item, InventorySystem.Items.Pickups.ItemPickupBase itemPickup)
